Configuration changes Pulled System.rb out to individual classes.
