+++ lib/objects/base_box.rb
@@ -1,0 +1,3 @@
+class Basebox
+  attr_accessor :name, :os, :distro, :vagrantbase, :url
+end
