+++ lib/objects/vulnerability.rb
@@ -3,11 +3,6 @@ require_relative('../constants.rb')
 class Vulnerability
     attr_accessor :type, :privilege, :access ,:puppets, :details, :ports, :name, :cve, :files, :scripts
 
-    def eql? other
-      # checks if type matches vulns.xml from scenario.xml
-      other.kind_of?(self.class) && @type == other.type
-    end
-
     def initialize(type='', privilege='', access='', puppets=[], details='', ports=[], platform ='', name='', cve='', files=[], scripts=[])
       @type = type
       @privilege = privilege
@@ -20,6 +15,11 @@ class Vulnerability
       @cve = cve
       @files = files
       @scripts = scripts
+    end
+
+    def eql? other
+      # checks if type matches vulns.xml from scenario.xml
+      other.kind_of?(self.class) && @type == other.type
     end
 
     def id
